src-python/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
aiosqlite>=0.19.0

# Search
meilisearch-python-sdk>=3.0.0

# Crawling
httpx>=0.26.0
aiofiles>=23.2.1
beautifulsoup4>=4.12.0
lxml>=5.1.0
trafilatura>=1.7.0
htmldate>=1.6.0

# Deduplication
simhash>=2.1.0

# Image embeddings (optional - for semantic image search)
sentence-transformers>=2.2.0
Pillow>=10.0.0
torch>=2.0.0

# Rate limiting
aiolimiter>=1.1.0

# External APIs
google-api-python-client>=2.114.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=8.2.0

# Scheduling (Windows)
pywin32>=306; sys_platform == 'win32'

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
